ml-service/requirements.txt
pydantic>=2.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# Heavy ML dependencies
--extra-index-url https://download.pytorch.org/whl/cpu
ultralytics>=8.0.196
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.2
pandas>=2.1.3
# onnxruntime==1.16.3
easyocr
timm
